+++ services/ingestion/etl/valor/xls/src/lib/transform.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { Project } from '../../../../_types/Project';
+import type { Project } from '../../../../../../../types/Project';
 import getCountryCode from '../../../../../helpers/getCountryCode';
 
 /*
