+++ types/Project.js
@@ -26,7 +26,7 @@ type BudgetItem = {
   currency: string,
   raw: string,
   value: number,
-  _original: {
+  _original?: {
     currency: string,
     raw: string,
     value: number,
